refactor(playground): extract AppShell nav data to navItems.ts + SEARCH_ITEMS
