+++ packages/playground/src/layout/AppShell/navItems.ts
@@ -1,0 +1,269 @@
+import {
+  AlertCircle,
+  Bell,
+  Building2,
+  LayoutDashboard,
+  KanbanSquare,
+  Users,
+  UserCog,
+  UserSquare,
+  MousePointer2,
+  CheckSquare,
+  CircleDot,
+  TextCursorInput,
+  FormInput,
+  ChevronsUpDown,
+  CalendarRange,
+  RectangleHorizontal,
+  Rows3,
+  Columns3,
+  Ruler,
+  LayoutGrid,
+  CircleUser,
+  Tag,
+  PanelTop,
+  PanelsTopLeft,
+  List,
+  ExternalLink,
+  FileText,
+  MessageSquare,
+  MessageSquareText,
+  AppWindow,
+  Minus,
+  MoreHorizontal,
+  Command,
+  KeyRound,
+  PanelLeft,
+  PanelRight,
+  ShieldCheck,
+  CalendarDays,
+  Clock,
+  Inbox,
+  Square,
+  Table as TableIcon,
+  TableProperties,
+  ListOrdered,
+  LayoutPanelLeft,
+  LayoutPanelTop,
+  Network,
+  Filter,
+  ListCollapse,
+  ListFilter,
+  MoveRight,
+  Sidebar,
+  ToggleRight,
+  GripVertical,
+  Heading,
+  Type,
+  Code as CodeIcon,
+  Activity,
+  LogIn,
+  LoaderCircle,
+  UploadCloud,
+  SlidersHorizontal,
+  Crop,
+  Image as ImageIcon,
+  Palette,
+  GalleryVerticalEnd,
+  Settings as SettingsIcon,
+  Fingerprint,
+  Hexagon,
+  TriangleAlert,
+  Compass,
+  Shapes,
+  SquareMousePointer,
+  SquarePen,
+  Group,
+  Columns2,
+  Component,
+  type LucideIcon,
+} from 'lucide-react';
+
+export interface NavItem {
+  to: string;
+  label: string;
+  icon: LucideIcon;
+  end: boolean;
+}
+
+export const mockupItems: NavItem[] = [
+  { to: '/mockups', label: 'Overview', icon: LayoutDashboard, end: true },
+  { to: '/mockups/dashboard', label: 'Dashboard', icon: LayoutDashboard, end: true },
+  { to: '/mockups/deals', label: 'Deals', icon: KanbanSquare, end: false },
+  { to: '/mockups/contacts', label: 'Contacts', icon: Users, end: false },
+  { to: '/mockups/members', label: 'Members', icon: UserCog, end: false },
+  { to: '/mockups/tenants', label: 'Tenants', icon: Building2, end: false },
+  { to: '/mockups/audit', label: 'Audit log', icon: Activity, end: false },
+  { to: '/mockups/system-settings', label: 'System settings', icon: SettingsIcon, end: false },
+  { to: '/mockups/custom-fields', label: 'Custom fields', icon: SquarePen, end: false },
+  { to: '/mockups/login', label: 'Login', icon: LogIn, end: false },
+  { to: '/mockups/404', label: 'Not found', icon: Compass, end: false },
+  { to: '/mockups/error', label: 'Error', icon: TriangleAlert, end: false },
+];
+
+export const componentOverview: NavItem = {
+  to: '/components',
+  label: 'Overview',
+  icon: Component,
+  end: true,
+};
+
+export const tokensReference: NavItem = {
+  to: '/components/tokens',
+  label: 'Design tokens',
+  icon: Palette,
+  end: false,
+};
+
+export const architectureReference: NavItem = {
+  to: '/components/architecture',
+  label: 'Architecture',
+  icon: Network,
+  end: false,
+};
+
+export const componentGroups: { heading: string; items: NavItem[] }[] = [
+  {
+    heading: 'Layout',
+    items: [
+      { to: '/components/app-layout', label: 'AppLayout', icon: PanelsTopLeft, end: false },
+      { to: '/components/stack', label: 'Stack', icon: Rows3, end: false },
+      { to: '/components/cluster', label: 'Cluster', icon: Columns3, end: false },
+      { to: '/components/constrain', label: 'Constrain', icon: Ruler, end: false },
+      { to: '/components/divider', label: 'Divider', icon: Minus, end: false },
+      { to: '/components/grid', label: 'Grid', icon: LayoutGrid, end: false },
+      { to: '/components/masonry', label: 'Masonry', icon: GalleryVerticalEnd, end: false },
+      { to: '/components/card', label: 'Card', icon: RectangleHorizontal, end: false },
+      { to: '/components/page-header', label: 'PageHeader', icon: LayoutPanelTop, end: false },
+      { to: '/components/page', label: 'Page', icon: FileText, end: false },
+      { to: '/components/screen', label: 'Screen', icon: AppWindow, end: false },
+    ],
+  },
+  {
+    heading: 'Forms',
+    items: [
+      { to: '/components/button', label: 'Button', icon: MousePointer2, end: false },
+      { to: '/components/button-group', label: 'ButtonGroup', icon: LayoutPanelLeft, end: false },
+      { to: '/components/social-button', label: 'SocialButton', icon: LogIn, end: false },
+      { to: '/components/checkbox', label: 'Checkbox', icon: CheckSquare, end: false },
+      { to: '/components/field', label: 'Field', icon: FormInput, end: false },
+      { to: '/components/form-row', label: 'FormRow', icon: Columns2, end: false },
+      { to: '/components/form-section', label: 'FormSection', icon: Group, end: false },
+      { to: '/components/color-picker', label: 'ColorPicker', icon: Palette, end: false },
+      { to: '/components/datepickers', label: 'Date pickers', icon: CalendarRange, end: false },
+      { to: '/components/file-upload', label: 'FileUpload', icon: UploadCloud, end: false },
+      { to: '/components/image-crop', label: 'ImageCrop', icon: Crop, end: false },
+      { to: '/components/input', label: 'Input', icon: TextCursorInput, end: false },
+      { to: '/components/password-input', label: 'PasswordInput', icon: KeyRound, end: false },
+      {
+        to: '/components/password-strength-meter',
+        label: 'PasswordStrengthMeter',
+        icon: ShieldCheck,
+        end: false,
+      },
+      { to: '/components/options-picker', label: 'OptionsPicker', icon: ListFilter, end: false },
+      { to: '/components/radio', label: 'Radio', icon: CircleDot, end: false },
+      { to: '/components/select', label: 'Select', icon: ChevronsUpDown, end: false },
+      { to: '/components/slider', label: 'Slider', icon: SlidersHorizontal, end: false },
+      { to: '/components/kanban', label: 'Kanban', icon: KanbanSquare, end: false },
+      { to: '/components/sortable', label: 'Sortable', icon: GripVertical, end: false },
+      { to: '/components/switch', label: 'Switch', icon: ToggleRight, end: false },
+      { to: '/components/textarea', label: 'Textarea', icon: MessageSquareText, end: false },
+      { to: '/components/timefield', label: 'TimeField', icon: Clock, end: false },
+    ],
+  },
+  {
+    heading: 'Display',
+    items: [
+      { to: '/components/avatar', label: 'Avatar', icon: CircleUser, end: false },
+      { to: '/components/badge', label: 'Badge', icon: Tag, end: false },
+      { to: '/components/brand-icon', label: 'BrandIcon', icon: Fingerprint, end: false },
+      { to: '/components/logo', label: 'Logo', icon: Hexagon, end: false },
+      { to: '/components/calendar', label: 'Calendar', icon: CalendarDays, end: false },
+      {
+        to: '/components/circular-progress',
+        label: 'CircularProgress',
+        icon: LoaderCircle,
+        end: false,
+      },
+      { to: '/components/code', label: 'Code', icon: CodeIcon, end: false },
+      { to: '/components/definition-list', label: 'DefinitionList', icon: List, end: false },
+      { to: '/components/empty-state', label: 'EmptyState', icon: Inbox, end: false },
+      { to: '/components/error-state', label: 'ErrorState', icon: TriangleAlert, end: false },
+      { to: '/components/filter-chip', label: 'FilterChip', icon: Filter, end: false },
+      { to: '/components/icon-tile', label: 'IconTile', icon: Shapes, end: false },
+      { to: '/components/image', label: 'Image', icon: ImageIcon, end: false },
+      { to: '/components/kbd', label: 'Kbd', icon: Command, end: false },
+      { to: '/components/pagination', label: 'Pagination', icon: ListOrdered, end: false },
+      { to: '/components/palette', label: 'Palette', icon: Palette, end: false },
+      { to: '/components/person-display', label: 'PersonDisplay', icon: UserSquare, end: false },
+      { to: '/components/progress', label: 'Progress', icon: Activity, end: false },
+      { to: '/components/skeleton', label: 'Skeleton', icon: Square, end: false },
+      { to: '/components/table', label: 'Table', icon: TableIcon, end: false },
+      { to: '/components/datatable', label: 'DataTable', icon: TableProperties, end: false },
+      { to: '/components/text', label: 'Text', icon: Type, end: false },
+      { to: '/components/title', label: 'Title', icon: Heading, end: false },
+    ],
+  },
+  {
+    heading: 'Feedback',
+    items: [
+      { to: '/components/alert', label: 'Alert', icon: AlertCircle, end: false },
+      { to: '/components/toast', label: 'Toast', icon: Bell, end: false },
+    ],
+  },
+  {
+    heading: 'Navigation',
+    items: [
+      { to: '/components/accordion', label: 'Accordion', icon: ListCollapse, end: false },
+      { to: '/components/breadcrumb', label: 'Breadcrumb', icon: MoveRight, end: false },
+      { to: '/components/link', label: 'Link', icon: ExternalLink, end: false },
+      { to: '/components/link-card', label: 'LinkCard', icon: SquareMousePointer, end: false },
+      { to: '/components/rail', label: 'Rail', icon: Sidebar, end: false },
+      { to: '/components/tabs', label: 'Tabs', icon: PanelTop, end: false },
+      { to: '/components/topbar', label: 'TopBar', icon: LayoutPanelTop, end: false },
+    ],
+  },
+  {
+    heading: 'Overlays',
+    items: [
+      { to: '/components/drawer', label: 'Drawer', icon: PanelRight, end: false },
+      { to: '/components/dropdown-menu', label: 'DropdownMenu', icon: MoreHorizontal, end: false },
+      { to: '/components/modal', label: 'Modal', icon: AppWindow, end: false },
+      { to: '/components/tooltip', label: 'Tooltip', icon: MessageSquare, end: false },
+      { to: '/components/popover', label: 'Popover', icon: PanelLeft, end: false },
+      {
+        to: '/components/confirmation-popover',
+        label: 'ConfirmationPopover',
+        icon: ShieldCheck,
+        end: false,
+      },
+    ],
+  },
+];
+
+/** One flat, searchable destination. */
+export interface SearchItem {
+  to: string;
+  label: string;
+  section: string;
+  icon: LucideIcon;
+}
+
+/**
+ * Flat jump-to list derived from the same arrays the rail renders, so a newly
+ * added component appears in search automatically (no second list to maintain).
+ * Order: references, then components (by group), then mockups.
+ */
+export const SEARCH_ITEMS: SearchItem[] = [
+  ...[componentOverview, tokensReference, architectureReference].map((i) => ({
+    to: i.to,
+    label: i.label,
+    section: 'Reference',
+    icon: i.icon,
+  })),
+  ...componentGroups.flatMap(({ heading, items }) =>
+    items.map((i) => ({ to: i.to, label: i.label, section: heading, icon: i.icon })),
+  ),
+  ...mockupItems.map((i) => ({ to: i.to, label: i.label, section: 'Mockups', icon: i.icon })),
+];
